// src/use-html-rewriter/html.schemas.ts
export const TemplateObjectSchema = z.object({
  html: z.array(z.string()),
  stylesheets: z.array(z.string()),
  scale: z.literal(Object.values(SCALE)),
  $: z.literal(TEMPLATE_OBJECT_IDENTIFIER),
})